+++ changelog.md
@@ -13,6 +13,7 @@ __[new features]__
 
 __[changes]__
 
+* Adjust search for newer version of Acrobat Reader 2020.
 * Fix failing search for newer versions of Blender LTS.
 * Starting with SeaMonkey 2.53.22, no more 32-bit builds for SeaMonkey are
   provided. Therefore the updater will now update 32-bit installations of